Build ASG end ELB via Ansible
The big "WTF" here is that the first time you create an ELB it will
fail. Re-run ansible again and it'll continue and finish setting things
up :(

Right now it will only create things tagged with an environment of
'test'. Set the environment variable of ENVIRONMENT=production to
something else if you want to prod.

The userdata `runcmd` will run on first boot only and will copy the
(read-only) DB from the EBS root volume onto the local SSD, adjust the
postgres config and then restart everything. When this is happening it
stops nginx so that it won't be in service with the ELB