Speed up AMI builds by building a base DB ami and the server off that
Since we need to import the 3GB addressbase table we don't want to have
to do that every time we make a code change -- so we build an AMI that
has that DB pre-imported. We have also got *a* version of the code in
this AMI to speed up the build for builds of the 'server' AMI: one of
the slowest bits there is waiting for the virtualenv to install/update
and if we've got a version of most modules already it's much quicker.

We didn't include the migrations for the addressbase table in the DB
dump we import, so we have to manually fix that up.