Build an extra AMI "layer" with imported DB.
Running the import with normal config takes quite a while (in the order
of 30-60minutes!) so we want to not have to do that on every code
change.

Just like we did with the addressbase import we pull down a PB dump from
an S3 bucket, and import it into postgres on the AMI. To make this
faster too I have a small script that tweaks the postgresql server
config to do less file IO -- this isn't safe on a running server but is
fine for us here.