Flush handlers in misc_fixes.yml
This is for the situation where the git pull changes something on the
frontend - eg template tags, and then there are errors for any requests
that happen between the git pull and misc_fixes being completed. This
should be minimal but is observable when we have a lot of traffix.

By flushing the handler queue the restart 'web frontend task'
will happen before running the various misc fixes. It shouldn't need to
happen after the misc fixes as they're operations on the DB not the
frontend.