Add a cron to import EONI data nightly and on reboot
This is necessary because EONI want us to stay in sync with a nightly
upload of their data to s3. This means we can't bake their data into an
image build, and instead need to import it when every instance is
started, and then nightly.

This can be turned off by commenting out the playbook import in
provision.yml.